# backend/openapi.snapshot.json
{
  "components": {
...
    "securitySchemes": {
      "accessToken": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "API operacional de Agency OS",
    "title": "Agency OS API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/agent/session/credential-grant": {
      "post": {
        "operationId": "AgentVaultController.grant",
        "parameters": [
          {
            "in": "header",
            "name": "x-device-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "AgentVault"
        ],
        "x-agency-device": true,
        "x-agency-permissions": [
          "vault.credential.issue"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "vault.credential.issue"
            ],
            "public": false,
            "roles": [
              "OPERADOR"
            ]
          },
          "actors": [
            "OPERADOR"
          ],
          "device": true,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/agent/session/credential-grant",
          "scope": "SELF",
          "shift": true
        },
        "x-agency-roles": [
          "OPERADOR"
        ],
        "x-agency-shift": true
      }
    },
    "/api/v1/agent/session/credential-redeem": {
      "post": {
        "operationId": "AgentVaultController.redeem",
        "parameters": [
          {
            "in": "header",
            "name": "x-device-token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "AgentVault"
        ],
        "x-agency-device": true,
        "x-agency-permissions": [
          "vault.credential.issue"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "vault.credential.issue"
            ],
            "public": false,
            "roles": [
              "OPERADOR"
            ]
          },
          "actors": [
            "OPERADOR"
          ],
          "device": true,
          "idempotency": "NOT_APPLICABLE",
          "method": "POST",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/agent/session/credential-redeem",
          "scope": "SELF",
          "shift": true
        },
        "x-agency-roles": [
          "OPERADOR"
        ],
        "x-agency-shift": true
      }
    },
...
    "/api/v1/profiles/{id}": {
      "get": {
        "operationId": "ProfilesController.get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Profiles"
        ],
        "x-agency-permissions": [
          "profiles.read"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "profiles.read"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "DIRECTOR_OPERATIVO",
            "COORDINADOR",
            "OPERADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "GET",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/profiles/{id}",
          "scope": "RESOURCE",
          "shift": false
        }
      },
      "patch": {
        "operationId": "ProfilesController.update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Profiles"
        ],
        "x-agency-permissions": [
          "profiles.update"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "profiles.update"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "DIRECTOR_OPERATIVO",
            "COORDINADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "PATCH",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/profiles/{id}",
          "scope": "RESOURCE",
          "shift": false
        }
      }
    },
...
    "/api/v1/profiles/{profileId}/credential": {
      "put": {
        "operationId": "VaultController.rotate",
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Vault"
        ],
        "x-agency-permissions": [
          "vault.rotate"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "vault.rotate"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN"
          ],
          "device": false,
          "idempotency": "NATURAL_KEY",
          "method": "PUT",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/profiles/{profileId}/credential",
          "scope": "RESOURCE",
          "shift": false
        }
      }
    },
    "/api/v1/profiles/{profileId}/credential/meta": {
      "get": {
        "operationId": "VaultController.meta",
        "parameters": [
          {
            "in": "path",
            "name": "profileId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "tags": [
          "Vault"
        ],
        "x-agency-permissions": [
          "vault.read_meta"
        ],
        "x-agency-policy": {
          "access": {
            "authenticated": true,
            "permissions": [
              "vault.read_meta"
            ],
            "public": false,
            "roles": []
          },
          "actors": [
            "ADMIN",
            "COORDINADOR",
            "OPERADOR"
          ],
          "device": false,
          "idempotency": "NOT_APPLICABLE",
          "method": "GET",
          "pagination": "NOT_APPLICABLE",
          "path": "/api/v1/profiles/{profileId}/credential/meta",
          "scope": "RESOURCE",
          "shift": false
        }
      }
    },
...
  },
  "servers": [],
...
}